MandelbrotRunnerFactory moved to api package. If a list with one entry is given, a single runner will be instantiated.
